refactor: extract account-only routes
