+++ lib/components/app/responsive-webapp.js
@@ -35,6 +35,8 @@ import MobileMain from '../mobile/main'
 import printRoutes from '../../util/webapp-print-routes'
 import tripPreviewRoutes from '../../util/webapp-trip-preview-routes'
 import webAppRoutes from '../../util/webapp-routes'
+// Can account routes be lazily-loaded?
+import accountRoutes from '../../util/account-routes'
 import withLoggedInUserSupport from '../user/with-logged-in-user-support'
 import withMap from '../map/with-map'
 
@@ -45,7 +47,12 @@ import SessionTimeout from './session-timeout'
 
 const { isMobile } = coreUtils.ui
 
-const routes = [...webAppRoutes, ...printRoutes, ...tripPreviewRoutes]
+const routes = [
+  ...webAppRoutes,
+  ...printRoutes,
+  ...tripPreviewRoutes,
+  ...accountRoutes
+]
 
 const MOBILE_CURRENT_POSITION_REFRESH_INTERVAL = 30000
